refactor(demos): migrate wave 1 (12 demos) to the shared style system
set-value, reset, clear, unset, modifiers, preprocess, records, radio,
select, handle-submit, inputs-to-submit and per-field-validation move
flat -> folder onto the shared registry.

Registry refinements driven by the real demos:
- light `pre` is the canonical JSON/state readout (theme-adaptive); the
  scattered dark terminal pres collapse onto it, leaving the dark console
  look to the `log` fragment for genuine append-logs.
- add an `.actions.mono` modifier for API-call button labels, a `stack`
  group for inner forms inside grid demos, and the shared `code` / `small`
  primitives the teaching demos lean on.
- `layout` now also works as a grid root (`class="demo layout"`); drop
  `panel`'s context-specific top margin.

Smoke green (63 passed, 4 skipped).
