+++ apps/site/docs-demos/clear/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
   import { useForm } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const form = useForm({
     schema: z.object({
@@ -18,7 +19,7 @@
 </script>
 
 <template>
-  <form @submit.prevent>
+  <form class="demo" @submit.prevent>
     <label>
       <span>Title <small v-if="form.fields.title.blank">(blank)</small></span>
       <input v-register="form.register('title')" />
@@ -30,16 +31,16 @@
       <small v-if="form.values.tags.length === 0">(blank: [])</small>
     </label>
 
-    <label class="check">
+    <label class="row compact">
       <input v-register="form.register('published')" type="checkbox" />
       Published
       <small v-if="form.fields.published.blank">(blank)</small>
     </label>
 
-    <div class="actions">
+    <div class="actions mono">
       <button type="button" @click="form.clear('title')">form.clear('title')</button>
       <button type="button" @click="form.clear('tags')">form.clear('tags')</button>
-      <button type="button" @click="form.clear()">form.clear() — whole form</button>
+      <button type="button" @click="form.clear()">form.clear() (whole form)</button>
     </div>
 
     <pre>{{ JSON.stringify(form.values, (_, v) => (v === undefined ? '(undefined)' : v), 2) }}</pre>
@@ -47,69 +48,16 @@
 </template>
 
 <style scoped>
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-    max-width: 30rem;
-  }
-  label {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-  }
-  label.check {
-    flex-direction: row;
-    align-items: center;
-    gap: 0.5rem;
-    font-weight: 400;
+  .row.compact {
     flex-wrap: wrap;
   }
-  label.check input {
-    margin: 0;
-  }
-  input {
-    padding: 0.5rem 0.75rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    font-size: 0.875rem;
-  }
   input[readonly] {
-    background: #f9fafb;
-    color: #6b7280;
+    background: var(--color-surface);
+    color: var(--color-fg-muted);
   }
   small {
     font-size: 0.75rem;
-    color: #2563eb;
+    color: var(--color-accent);
     font-weight: 500;
-  }
-  .actions {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.4rem;
-  }
-  .actions button {
-    padding: 0.35rem 0.7rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    background: #fff;
-    font-size: 0.75rem;
-    font-family: ui-monospace, monospace;
-    cursor: pointer;
-  }
-  .actions button:hover {
-    background: #f3f4f6;
-  }
-  pre {
-    background: #f9fafb;
-    border: 1px solid #e5e7eb;
-    border-radius: 0.375rem;
-    padding: 0.5rem 0.75rem;
-    font-size: 0.75rem;
-    font-family: ui-monospace, monospace;
-    color: #111827;
-    margin: 0;
   }
 </style>
