+++ apps/site/docs-demos/clear/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "input", "row", "actions", "pre"] }
