+++ apps/site/docs-demos/handle-submit/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
   import { useForm } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const form = useForm({
     schema: z.object({
@@ -22,13 +23,13 @@
 </script>
 
 <template>
-  <form @submit.prevent="onSubmit">
+  <form class="demo" @submit.prevent="onSubmit">
     <label>
       Email
       <input v-register="form.register('email')" autocomplete="email" />
       <em v-if="form.fields.email.showErrors">{{ form.fields.email.firstError?.message }}</em>
     </label>
-    <label class="checkbox">
+    <label class="row">
       <input v-register="form.register('terms')" type="checkbox" />
       I accept the terms of service
       <em v-if="form.fields.terms.showErrors">{{ form.fields.terms.firstError?.message }}</em>
@@ -38,60 +39,3 @@
     </button>
   </form>
 </template>
-
-<style scoped>
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-    max-width: 24rem;
-  }
-  label {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-  }
-  label.checkbox {
-    flex-direction: row;
-    align-items: center;
-    gap: 0.5rem;
-    flex-wrap: wrap;
-  }
-  input {
-    padding: 0.5rem 0.75rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    font-size: 0.875rem;
-  }
-  label.checkbox input {
-    padding: 0;
-    width: auto;
-  }
-  input:focus {
-    outline: 2px solid #2563eb;
-    outline-offset: -1px;
-  }
-  em {
-    color: #dc2626;
-    font-size: 0.8125rem;
-    font-style: normal;
-    font-weight: 400;
-  }
-  button {
-    margin-top: 0.25rem;
-    padding: 0.625rem 1rem;
-    background: #2563eb;
-    color: white;
-    border: none;
-    border-radius: 0.375rem;
-    font-size: 0.875rem;
-    font-weight: 600;
-    cursor: pointer;
-  }
-  button:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-  }
-</style>
