+++ apps/site/docs-demos/handle-submit/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "input", "error", "row", "button"] }
