+++ apps/site/docs-demos/inputs-to-submit/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "input", "error", "row", "button"] }
