+++ apps/site/docs-demos/modifiers/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
   import { useForm } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const form = useForm({
     schema: z.object({
@@ -13,7 +14,7 @@
 </script>
 
 <template>
-  <form @submit.prevent>
+  <form class="demo" @submit.prevent>
     <label>
       <span><code>.lazy</code>: writes on change/blur, not on every keystroke</span>
       <input v-register.lazy="form.register('lazyName')" placeholder="Type, then blur" />
@@ -44,47 +45,3 @@
     </label>
   </form>
 </template>
-
-<style scoped>
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-    max-width: 28rem;
-  }
-  label {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    font-size: 0.8125rem;
-    font-weight: 500;
-  }
-  code {
-    font-family: ui-monospace, monospace;
-    background: #f3f4f6;
-    padding: 0.05rem 0.3rem;
-    border-radius: 0.25rem;
-    font-weight: 600;
-  }
-  input {
-    padding: 0.5rem 0.75rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    font-size: 0.875rem;
-  }
-  input:focus {
-    outline: 2px solid #2563eb;
-    outline-offset: -1px;
-  }
-  small {
-    font-size: 0.75rem;
-    font-weight: 400;
-    color: #374151;
-    font-family: ui-monospace, monospace;
-  }
-  em {
-    color: #2563eb;
-    font-style: normal;
-    font-weight: 500;
-  }
-</style>
