+++ apps/site/docs-demos/modifiers/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "code", "input", "small"] }
