+++ apps/site/docs-demos/per-field-validation/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "input", "error"] }
