+++ apps/site/docs-demos/preprocess/App.vue
@@ -2,6 +2,7 @@
   import { computed, ref } from 'vue'
   import { useForm } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const normalize = (v: unknown): unknown => (typeof v === 'string' ? v.trim().toLowerCase() : v)
 
@@ -23,8 +24,8 @@
 </script>
 
 <template>
-  <div class="layout">
-    <form @submit.prevent="onSubmit">
+  <div class="demo layout">
+    <form class="stack" @submit.prevent="onSubmit">
       <label>
         Email
         <input
@@ -73,64 +74,10 @@
 </template>
 
 <style scoped>
-  .layout {
-    display: grid;
-    grid-template-columns: 1fr;
-    gap: 1.25rem;
-  }
   @media (min-width: 760px) {
-    .layout {
+    .demo.layout {
       grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
     }
-  }
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.625rem;
-  }
-  label {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    font-size: 0.8125rem;
-    color: #374151;
-  }
-  input {
-    padding: 0.5rem 0.625rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    font-size: 0.875rem;
-    font-family: inherit;
-  }
-  input:focus {
-    outline: 2px solid #2563eb;
-    outline-offset: -1px;
-  }
-  button {
-    align-self: flex-start;
-    padding: 0.5rem 0.875rem;
-    border-radius: 0.375rem;
-    border: 1px solid #2563eb;
-    background: #2563eb;
-    color: white;
-    font-size: 0.8125rem;
-    font-weight: 500;
-    cursor: pointer;
-    font-family: inherit;
-  }
-  button:hover {
-    background: #1d4ed8;
-  }
-  .error {
-    margin: 0;
-    color: #b91c1c;
-    font-size: 0.8125rem;
-    font-weight: 500;
-  }
-  .hint {
-    margin: 0;
-    color: #6b7280;
-    font-size: 0.75rem;
   }
   section {
     display: flex;
@@ -145,23 +92,6 @@
   section p {
     margin: 0;
     font-size: 0.75rem;
-    color: #6b7280;
-  }
-  pre {
-    margin: 0;
-    padding: 0.5rem 0.625rem;
-    background: #0f172a;
-    color: #a5f3fc;
-    border-radius: 0.375rem;
-    font-size: 0.75rem;
-    font-family: ui-monospace, monospace;
-    overflow: auto;
-  }
-  code {
-    font-family: ui-monospace, monospace;
-    background: #f3f4f6;
-    padding: 0.05rem 0.3rem;
-    border-radius: 0.25rem;
-    font-size: 0.75rem;
+    color: var(--color-fg-muted);
   }
 </style>
