+++ apps/site/docs-demos/preprocess/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["layout", "stack", "label", "input", "button", "error", "hint", "code", "pre"] }
