+++ apps/site/docs-demos/radio/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
   import { useForm } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const form = useForm({
     schema: z.object({
@@ -12,7 +13,7 @@
 </script>
 
 <template>
-  <form @submit.prevent>
+  <form class="demo" @submit.prevent>
     <fieldset>
       <legend>Plan</legend>
       <label class="row">
@@ -34,48 +35,3 @@
     }}</pre>
   </form>
 </template>
-
-<style scoped>
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-    max-width: 26rem;
-  }
-  fieldset {
-    display: flex;
-    flex-direction: column;
-    gap: 0.4rem;
-    border: 1px solid #e5e7eb;
-    border-radius: 0.375rem;
-    padding: 0.6rem 0.75rem;
-  }
-  legend {
-    padding: 0 0.4rem;
-    font-size: 0.75rem;
-    font-weight: 600;
-    color: #6b7280;
-    text-transform: uppercase;
-    letter-spacing: 0.05em;
-  }
-  .row {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-size: 0.875rem;
-    font-weight: 400;
-  }
-  input {
-    margin: 0;
-  }
-  pre {
-    background: #f9fafb;
-    border: 1px solid #e5e7eb;
-    border-radius: 0.375rem;
-    padding: 0.5rem 0.75rem;
-    font-size: 0.75rem;
-    font-family: ui-monospace, monospace;
-    color: #111827;
-    margin: 0;
-  }
-</style>
