+++ apps/site/docs-demos/records/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["fieldset", "row", "code", "hint", "pre"] }
