+++ apps/site/docs-demos/select/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "select", "pre"] }
