+++ apps/site/docs-demos/set-value/styles.json
@@ -1,0 +1,1 @@
+{ "with": ["label", "input", "actions", "pre"] }
