+++ apps/site/docs-demos/unset/App.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
   import { useForm, unset } from 'attaform/zod'
   import { z } from 'zod'
+  import './styles.css'
 
   const schema = z.object({
     email: z.string(),
@@ -19,7 +20,7 @@
 </script>
 
 <template>
-  <form @submit.prevent>
+  <form class="demo" @submit.prevent>
     <label>
       <span>Email (primitive leaf)</span>
       <input v-register="form.register('email')" />
@@ -37,7 +38,7 @@
       </label>
     </fieldset>
 
-    <div class="actions">
+    <div class="actions mono">
       <button type="button" @click="form.setValue('email', unset)">
         setValue('email', unset)
       </button>
@@ -47,7 +48,7 @@
       <button type="button" @click="form.reset()">reset()</button>
     </div>
 
-    <div class="panel">
+    <div class="kv">
       <p>
         <code>form.values</code> =
         <em>{{ JSON.stringify(form.values, null, 2) }}</em>
@@ -65,77 +66,23 @@
 </template>
 
 <style scoped>
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.875rem;
-    max-width: 32rem;
-  }
-  label {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-  }
-  fieldset {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    border: 1px solid #d1d5db;
-    border-radius: 0.5rem;
-    padding: 0.75rem;
-  }
-  legend {
-    font-size: 0.8125rem;
-    font-weight: 600;
-    color: #374151;
-    padding: 0 0.375rem;
-  }
-  input {
-    padding: 0.5rem 0.75rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    font-size: 0.875rem;
-  }
-  input:focus {
-    outline: 2px solid #2563eb;
-    outline-offset: -1px;
-  }
-  .actions {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.4rem;
-  }
-  .actions button {
-    padding: 0.35rem 0.7rem;
-    border-radius: 0.375rem;
-    border: 1px solid #d1d5db;
-    background: #fff;
-    font-size: 0.75rem;
-    font-family: ui-monospace, monospace;
-    cursor: pointer;
-  }
-  .actions button:hover {
-    background: #f3f4f6;
-  }
-  .panel {
-    background: #f9fafb;
-    border: 1px solid #e5e7eb;
+  .kv {
+    background: var(--color-surface);
+    border: 1px solid var(--color-border);
     border-radius: 0.375rem;
     padding: 0.5rem 0.75rem;
     font-size: 0.8125rem;
     font-family: ui-monospace, monospace;
   }
-  .panel p {
+  .kv p {
     margin: 0.2rem 0;
     white-space: pre-wrap;
   }
-  code {
-    color: #6b7280;
+  .kv code {
+    color: var(--color-fg-muted);
   }
-  em {
-    color: #2563eb;
+  .kv em {
+    color: var(--color-accent);
     font-style: normal;
     font-weight: 500;
   }
